Fix publish date issue and improve TableOfContents
ACTION REQUIRED (see "Updating" below)

When support for .Params.publishdate was introduced in the last few
commits it caused an issue for posts only using .Params.date. This
commit resolves that issue by falling back to .Date when .PublishDate
is not available for a post.

TableOfContents no longer appears by default and is an opt-in
feature.

Fixed a few other cosmetic bugs and added more documentation.

Updating: Remove `notoc` in any posts. It's no longer used. To display
a table of contents use `toc` instead. See README for usage.