Fix bugs in publish date display
When publish dates were fixed to prevent "0001-01-01" from appearing
in the byline I introduced a bug whereby a pages last modified date
would be shown in the Byline. This fixes that issue so the Byline always
shows the date an item was originally published.

Additionally, instead of showing "Updated" for every item in post
listings we now display the "Published" date or, if updated, the
"Updated" date instead, regardless of use of .Params.publishdate.