feat($archetypes): turn archetypes into go templates
hugo 0.24 and newer allow archetypes to be go templates. this feature updates the archetype markdown files so they are now go templates.

BREAKING CHANGE: Minimum Hugo version set 0.24, 0.16 - 0.23 must modify archetypes

to upgrade add `date` and `title` to custom archetypes. if using a version of hugo prior to 0.24 the custom templates should remove `title` and `date` and not use go templating logic in the archetype files